backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.34.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-cleanup==8.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
oauthlib==3.2.2
packaging==23.1
phonenumberslite==8.13.20
Pillow==10.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
dj-rest-auth==5.0.1
urllib3==1.26.16
# django-redis==5.4.0
gevent==23.9.1
